[ BE ] (perf) ItemService 클래스 updateItem() 메서드 매개변수 수정
[ BE ] (perf) ItemService 클래스 updateItem() 메서드 매개변수 수정

상품 수량의 타입을 int 로 수정해주었습니다.

+++ 4 WEEK/item-service/src/main/java/hello/itemservice/item/ItemServiceImpl.java
@@ -18,7 +18,7 @@ public class ItemServiceImpl implements ItemService{
     }
 
     @Override
-    public void updateItem(Item item, String itemName, double itemPrice, long itemQuantity) {
+    public void updateItem(Item item, String itemName, double itemPrice, int itemQuantity) {
         item.setName(itemName);
         item.setPrice(itemPrice);
         item.setQuantity(itemQuantity);
